feat: ingest stronger from private-graph
circular deps elimination in the graph code introduced duplication in
the code which i don't want to eliminate manually. private-graph has the
same code so the source shouldn't matter.
